pyproject.toml
[project]
name = "my-first-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.96.0",
    "django>=6.0.4",
    "django-ninja>=1.6.2",
    "pydantic>=2.13.3",
    "python-decouple>=3.8",
]

[dependency-groups]
dev = [
    "ruff>=0.15.11",
]
[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "DJ"] # General, Imports, and Django-specific rules